+++ docs/CHANGELOG.md
@@ -10,6 +10,7 @@ Adds a custom model.
 + All dyed amethyst clusters and buds must be harvested with silk touch.
 + Dyed amethyst clusters broken without silk touch now drop shards.
 + Added Parkson Inc. Company-Issued Hard hat.
++ Added Blowout Microphone model.
 
 # Release 2.0.0